Several fixes to OPCODE plugin to make it compile.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@10997 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/collide/opcode/OPC_HPoint.h
@@ -62,7 +62,7 @@
 		//! Computes square magnitude
 		inline_	float		SquareMagnitude()					const		{ return x*x + y*y + z*z + w*w;											}
 		//! Computes magnitude
-		inline_	float		Magnitude()							const		{ return sqrtf(x*x + y*y + z*z + w*w);									}
+		inline_	float		Magnitude()							const		{ return sqrt(x*x + y*y + z*z + w*w);									}
 
 		//! Normalize the vector
 		inline_	HPoint&		Normalize()
